Use Cocoa integer types to facilitate 64-bit transition

+++ keycastr/KCAppController.h
@@ -52,7 +52,7 @@
 	IBOutlet NSMenuItem* dockShortcutItem;
 	BOOL _isCapturing;
 	BOOL _allowToggle;
-	int _startupIconPreference;
+	NSInteger _startupIconPreference;
     KCKeyboardTap* keyboardTap;
 }
